[IMP] website_slides: quickly lint tests
Purpose is to make a quick reordering / linting of tests while passing by.

Task-2607416
