+++ addons/website_slides/tests/__init__.py
@@ -3,9 +3,10 @@
 
 from . import common
 from . import test_embed_detection
-from . import test_karma
-from . import test_resources
+from . import test_gamification_karma
 from . import test_security
-from . import test_slide_utils
+from . import test_slide_channel
+from . import test_slide_resource
+from . import test_slide_slide
 from . import test_statistics
 from . import test_ui_wslides
