+++ addons/website_slides/tests/test_gamification_karma.py
@@ -7,7 +7,7 @@ from odoo.tests.common import users
 from odoo.tools import mute_logger
 
 
-@tagged('functional')
+@tagged('gamification')
 class TestKarmaGain(common.SlidesCase):
 
     def setUp(self):
@@ -25,20 +25,20 @@ class TestKarmaGain(common.SlidesCase):
             'karma_gen_channel_rank': 10,
         })
 
-        self.slide_2_0 = self.env['slide.slide'].with_user(self.user_officer).create({
-            'name': 'How to travel through space and time',
-            'channel_id': self.channel_2.id,
-            'slide_category': 'document',
-            'is_published': True,
-            'completion_time': 2.0,
-        })
-        self.slide_2_1 = self.env['slide.slide'].with_user(self.user_officer).create({
-            'name': 'How to duplicate yourself',
-            'channel_id': self.channel_2.id,
-            'slide_category': 'document',
-            'is_published': True,
-            'completion_time': 2.0,
-        })
+        self.slide_2_0, self.slide_2_1 = self.env['slide.slide'].with_user(self.user_officer).create([
+            {'name': 'How to travel through space and time',
+             'channel_id': self.channel_2.id,
+             'slide_category': 'document',
+             'is_published': True,
+             'completion_time': 2.0,
+            },
+            {'name': 'How to duplicate yourself',
+             'channel_id': self.channel_2.id,
+             'slide_category': 'document',
+             'is_published': True,
+             'completion_time': 2.0,
+            }
+        ])
 
     @mute_logger('odoo.models')
     @users('user_emp', 'user_portal', 'user_officer')
